Added previous turn board state as a Virtual River variable
System now also keeps track of the total costs and turn costs and updates the total costs only after a turn is ended.

This means that a new update is always compared to the previous board state, as opposed to the previous update. Therefore, new updates within the same turn are possible. However, this needs to be tested. Possibly, previous grids need to also be stored.